feat
Added existence checks before PUT and DELETE methods

+++ controllers/PetController.js
@@ -40,16 +40,27 @@ module.exports = class PetController {
       date_of_birth,
     };
 
-    await Pet.update(userData, { where: { id: petId, TutorId: tutorId } });
-    res.json(Pet);
+    const pet = await Pet.findOne({where: { id: petId, TutorId: tutorId }})
+    if(!pet){
+      res.json('Pet not found')
+    }else {
+      await Pet.update(userData, { where: { id: petId, TutorId: tutorId } });
+      res.json(Pet);
+    }
   }
 
   static async deletePet(req, res) {
     const tutorId = req.params.tutorId;
     const petId = req.params.petId;
 
-    await Pet.destroy({ where: { id: petId, TutorId: tutorId } });
-
-    res.json("status code 204");
+    
+    const pet = await Pet.findOne({where: { id: petId, TutorId: tutorId }})
+    if(!pet){
+      res.json('Pet not found')
+    }else {
+      await Pet.destroy({ where: { id: petId, TutorId: tutorId } });
+  
+      res.json("status code 204");
+    }
   }
 };
